fix: reverted to old game logic, removing game engine
- Removed engine/ directory completely
- Removed Zustand stores (gameStore, userStore)
- Removed all engine tests
- Reverted multiplayer page to proven working version
- All Turso/PWA/layout changes still intact
